requirements.txt
cairosvg==2.7.1
Pillow==10.3.0